requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
rapidfuzz==3.5.2
gspread==6.1.2
google-auth==2.35.0
